Add getLabelText util
