+++ services/ui-src/src/components/Upload/index.tsx
@@ -7,7 +7,7 @@ import { useController, useFormContext } from "react-hook-form";
 import { Storage } from "aws-amplify";
 import { saveAs } from "file-saver";
 import { useUser } from "hooks/authHooks";
-import { createSafeS3Key } from "utils/cleanString";
+import { createSafeS3Key } from "utils";
 
 interface IUploadProps {
   maxSize?: number;
