--- services/ui-src/src/hooks/useLabelText.tsx
@@ -1,25 +1,0 @@
-import { useLocation } from "react-router-dom";
-
-export const useLabelText = () => {
-  const { pathname } = useLocation();
-  const params = pathname.split("/");
-  const year = params[2];
-  const measure = params[4].replace("-", "");
-
-  const data = require(`../measures/${year}/${measure}/data`);
-
-  return {
-    ...data.qualifierLabelsAndText,
-    ...data.categoryLabelsAndText,
-  };
-};
-
-/**
- *
- * qualifiers: [{
- * id: 'somestring',
- * label: 'some string',
- *
- *
- * }]
- */
