+++ services/ui-src/src/measures/2021/AMBCH/index.tsx
@@ -3,7 +3,7 @@ import * as PMD from "./data";
 import * as QMR from "components";
 import { FormData } from "./types";
 import { getPerfMeasureRateArray } from "measures/2021/globalValidations";
-import { positiveNumbersWithMaxDecimalPlaces } from "utils/numberInputMasks";
+import { positiveNumbersWithMaxDecimalPlaces } from "utils";
 import { useEffect } from "react";
 import { useFormContext } from "react-hook-form";
 import { validationFunctions } from "./validation";
