+++ services/ui-src/src/measures/2021/CommonQuestions/PerformanceMeasure/index.tsx
@@ -5,8 +5,7 @@ import * as Types from "../types";
 import * as DC from "dataConstants";
 import { PerformanceMeasureData } from "./data";
 import { useWatch } from "react-hook-form";
-import { cleanString } from "utils/cleanString";
-import { useLabelText } from "hooks/useLabelText";
+import { cleanString, getLabelText } from "utils";
 import { ndrFormula } from "types";
 
 interface Props {
@@ -59,7 +58,7 @@ const CategoryNdrSets = ({
   customRateLabel,
 }: NdrSetProps) => {
   const register = useCustomRegister();
-  const labelText = useLabelText();
+  const labelText = getLabelText();
 
   return (
     <>
