+++ services/ui-src/src/measures/2022/globalValidations/ComplexValidations/ComplexValidateNDRTotals/index.tsx
@@ -1,4 +1,4 @@
-import { cleanString } from "utils/cleanString";
+import { cleanString } from "utils";
 
 interface NDRforumla {
   numerator: number;
